+++ packages/cloud-toolkit/src/command.ts
@@ -11,7 +11,7 @@ import DistPathManager from './services/dist-path-manager';
 import { getEnvironmentVariables } from './services/environment-variables';
 import EventSubscriber from './services/event-subscriber';
 import HttpServer from './services/http-server';
-import { EnvironmentVariables } from './types';
+import { EnvironmentVariables, Logger } from './types';
 
 configDotenv();
 
@@ -23,12 +23,20 @@ const buildEventSubscriber = (envs: EnvironmentVariables): EventSubscriber => {
   return new EventSubscriber(envs.FOREST_SUBSCRIPTION_URL, envs.FOREST_AUTH_TOKEN);
 };
 
+const logger: Logger = {
+  spinner: ora(),
+  // eslint-disable-next-line no-console
+  log: (text?: string) => console.log(text),
+  // eslint-disable-next-line no-console
+  error: (text?: string) => console.error(text),
+};
+
 const command = makeCommands({
   getEnvironmentVariables,
   buildHttpServer,
   buildEventSubscriber,
   login,
-  spinner: ora(),
+  logger,
   bootstrapPathManager: new BootstrapPathManager(os.tmpdir(), os.homedir()),
   distPathManager: new DistPathManager(),
 });